feat(netsim): factor packet and related definitions
I want to add more packages for additional functionality like
delays, packet filtering, etc. The first step is to ensure the
definition of packet and related definitions, e.g., NetworkDevice
are shared, so other packages cna use them.
